One exp per (centre, exponent) group in the grid evaluators; NNLS reject loop; centre-multipole golden
The wfn primitive order repeats every exponent once per Cartesian component of
its shell, so compute_dens_cartesian, computeRhoELI, computeGrad and
computeELIGrad evaluated exp(-a r^2) 3/6/10 times per shell per grid point.
WFN::build_exp_groups (built with the primitive-major coefficient cache, same
invalidation, also after change_exponent/change_center) numbers the distinct
(centre, exponent) pairs centre by centre; WFN::exp_table fills one exp per
group at a point, 0 below constants::exp_cutoff, and skips a whole centre when
its most diffuse primitive is already below the cutoff - the same test the
per-primitive loop made, so the evaluators, which read exps[group[j]] in the
old primitive order, give byte-identical results. FLOWOFFICE, 8 threads,
-eli_analysis with the full basin pipeline: water CASSCF 0.05 A 4.5 -> 3.6 s,
sucrose 0.1 A 528 -> 497 s, logs identical after dropping the timing lines.
WfnOpsTests.ExpTableMatchesPerPrimitiveExp checks the group count, every
primitive's table value and the whole-centre skip.

nnls: the two column-reject paths lost scipy's skip = True in the port, so the
next pass recomputed w and handed the rejected column its gradient back; a
dependent column whose gradient rounds to +eps (Linux, both CI jobs) therefore
never terminated and CellMathTests.NnlsIterationLimitAndDependentColumn ran
into the 1500 s ctest timeout. The reject now zeroes the positional w[iz] the
argmax reads (w[j] used the column index) and skips the recompute.

tests/epoxide_gbw/ri_fit_multipoles_centre.good: regenerated, the l=0 rows now
carry the fitted monopoles that the review fix fills (they sum to 23.98 e for
C2H4O); the golden had been recorded with the 0.000000 rows.

CodeQL: (lmax + 1) * (lmax + 1) widened to size_t before the vec size in
gaussian_atom.cpp and GridManager.cpp.

Full suite 1015 passed / 8 skipped / 0 failed (385 s, 8 threads).

+++ Src/core/GridManager.cpp
@@ -709,7 +709,7 @@ vec2 GridManager::calculatePartitionedMultipoles(const WFN &wave, const int lmax
 	vec3 *grid = needs_helper_grids_ ? &grid_data_.helper_grids : &grid_data_.atomic_grids;
 	const int num_atoms = (int)grid->size();
 	const GridData::GridIndex weight_index = partitionWeightIndex();
-	vec2 moments(num_atoms, vec((lmax + 1) * (lmax + 1), 0.0));
+	vec2 moments(num_atoms, vec((size_t)(lmax + 1) * (lmax + 1), 0.0));
 #pragma omp parallel for schedule(dynamic)
 	for (int atom = 0; atom < num_atoms; atom++) {
 		const vec2 &atomic_grid = (*grid)[atom];
